BgpAdvertisementGroup data model class (#2786)

+++ projects/batfish/src/main/java/org/batfish/grammar/flatjuniper/ConfigurationBuilder.java
@@ -5691,9 +5691,9 @@ public class ConfigurationBuilder extends FlatJuniperParserBaseListener {
 
   private long toCommunityLong(Ec_literalContext ctx) {
     String[] parts = ctx.getText().split(":");
-    int part1 = Integer.parseInt(parts[0]);
+    long part1 = Long.parseLong(parts[0]);
     long part2 = Long.parseLong(parts[1]);
-    int part3 = Integer.parseInt(parts[2]);
+    long part3 = Long.parseLong(parts[2]);
     ExtendedCommunity c = new ExtendedCommunity(part1, part2, part3);
     return c.asLong();
   }
